requirements.txt
dotenv==0.9.9
discord.py==2.6.4
Pillow>=10.0.0